Change from "marketing" to "communications"
In the process of documenting this role (PR soon!), it became clear that this is not a "marketing" role, but instead a communications coordinator position between CNCF, media, PM, and the release team.  There is an assumption that the CNCF will have a person dedicated to their marketing efforts already.