add release diagrams
I made these for my KubeCon EU 2018 presentation and they seem to be
useful in giving a visual idea of the overall flow for people.  I'm not
sure if there's a more portable and reusable way to export a Google
Drawing than just dumping a PNG file, if so I'm open to suggestions.

Signed-off-by: Tim Pepper <tpepper@vmware.com>