1.12 release timeline
In general the timeline across the calendar quarter for the release is
similar to past releases.  A few points to highlight:

Shifts in 1.11 timeline especially around code slush and freeze are
carrying forward into 1.12.  This is now the schedule intent, but if CI
signal is not green as we get toward the intended sluch and freeze we
may need to elongate them by starting them earlier.  This would be
communicated early and broadly if it is an incoming risk to the release
stability mid-cycle.

During the feature definition phase we will re-emphasize past intent
to have initial docs drafts as well as test cases arriving in
conjunction with the code at or ahead of code freeze.

In the final days of the cycle we lack the time (twelve week cycle
versus desire for shorter freeze versus wanting freeze as late as
possible) to dedicate more time for CI signal from long running
tests during the "thaw" period when cherry picks from master present
instability risk.  We may postponing more incoming cherry picks to
the point release depending on risk assessments.

Signed-off-by: Tim Pepper <tpepper@vmware.com>