process docs: note bug triage should track recent commits in burndown
We've had surprise commits merge late and without release team
visibility.  The milestone maintainers list is large and we have a
pragmatically light process favoring code velocity.  Nevertheless the
release team needs to have context of what's changing the release and
weigh stability risks of the various things in flight.

Signed-off-by: Tim Pepper <tpepper@vmware.com>