update alpha cadence
As it's been explained to me, the first alpha (alpha.0) is autocreated
at the time of the prior release's branch creation.  For...reasons?

That leaves the second alpha (alpha.1) as the first point where the
current release team builds an alpha release and the new release branch
manager gets to work out any process and permissions kinks.

Looking back it appears there's been legacy copy/pasted info implying
the first alpha was .1 and happened in week 2, but that hasn't been the
case.  This commit brings the calendar for 1.12 in line with what
happened on prior releases and is happening for 1.12.

Signed-off-by: Tim Pepper <tpepper@vmware.com>