move process ephemera to community repo devel guide
This document describes how the community targets issues and PRs
to a milestone release.  As such I think it fits best in the developer
guide.  A portion of the information on escalating issues relates
to the release team, but is already covered in the release-team
role handbooks and that escalation process is good to have visible
to developers broadly so they understand that they're liable to be
pinged by the release team for clarifying information.

Similarly the images showing the release process graphically then
would better sit in the developer guide also.

See also https://github.com/kubernetes/community/pull/2566

Signed-off-by: Tim Pepper <tpepper@vmware.com>