update 1.12 urls and process info
Somehow with the release process docs restructure it appears the
ephemera/issues.md link got broken.  Rather than just fix the link,
a separate commit is moving that document to the k8s developer guide
in the community repo.  That leaves a link needing cleaned up here.

Older releases (1.10 and 1.11) are left with the historical url which could
be looked up in git history if one desired that historical process context.

Part of moving the old issues.md document to k/community was also
streamlining the description there.  A few bits of wording here are brought
in line with the new release.md document linked.

Signed-off-by: Tim Pepper <tpepper@vmware.com>