Add upgrade job version guide
I have been hearing lots of confusion over whether or not a given fix
has made it into an upgrade job. Often times this is of the form "I
think the fix got merged, but it didn't make it into the release for
some reason"

This guide shows how to quickly determine what versions of kubernetes a
job is upgrading FROM and TO. Very few jobs actually use releases. If
you don't trust the job, the guide shows how you can get a k/k commit
SHA out of the build log, and use that with github's ui to see exactly
which commits/merges made it in for that job.