Remove references to code slush from bugtriage playbook
This was the toughest to change, I feel like it needs more
of a rewrite from the Bug Triage folks. Tons of text, and
more phases than the actual release cycle. It also talked
about Burndown solely during the last week, which isn't
quite true.

In the context of duties changing as phases change, I
changed "Slush" to "Burndown", and also changed
"Release Burndown" to "Code Thaw"