patch release handbook: add workflow reference example
The order of operations in the release workflow is a mix of complicated,
error-prone command invocations, and manual synchronization check
points.  A summary of these is needed to make the required actions more
clear.

Also adds a table of contents as intradoc links makes eaiser navigating to the
appropriate info a different points in a process that spans days or even
weeks.

Signed-off-by: Tim Pepper <tpepper@vmware.com>