[WIP] patch release handbook: propose abstracted workflow
For a concrete discussion on tools improvement, I want to show what
I'm imagining it might look like to establish a wrapping of the
detailed commands which do build and publish.

My goal is to allow parallel back end implementations and
proof-of-concept comparisons.  If the top level flow is abstracted
and callable by automation, without those tools getting coupled to
the internal implementation, we are in a more flexible and testable
place and enable more to be done through automation.