document Patch Release Team spreadsheet process
GitHub doesn't enable concise tracking of the evolving state of PRs
as they shuffle along towards meeting merge criteria on a branch.
During the past 9 months we've evolved a process for this in the
patch release team so we can collaborate on triaging inbound cherry
picks and driving consensus on whether to accept them on a branch.
This new documentation attempts to capture the current process.  It
reads a lot more complicated than it is, although better tooling
would nevertheless be very welcome.

Signed-off-by: Tim Pepper <tpepper@vmware.com>