patch release schedule: active and non-active branch description
Between this document and the k8s.io website it remains less than fully
clear what are the current active branches, especially when 1.19 comes
out.  The website is getting some clarification in the 1.19 branch at
https://github.com/kubernetes/website/pull/21928#issuecomment-649688854

A follow change here for 1.19 and newer would be to add additional
mention of through when we're expecting a currently active branch to go
EOL and non-active.

Signed-off-by: Tim Pepper <tpepper@vmware.com>