schedule: update for Aug. 2020 and 1.16 heading toward EOL
Regular monthly update forcasting the coming patch release schedule.

NOTE: With an August 2020 release of Kubernetes 1.19, we trigger
end-of-support for the N-3 branch 1.16.  As of August, instead of
the typical nine months, the 1.16 branch will have had almost one
year of patch backports due to the elongation of the 1.19 release
cycle.  Lacking some critical-urgent cherry pick content pending
in August and meriting one additional release in September, it feels
reasonable to let this branch go.

Signed-off-by: Tim Pepper <tpepper@vmware.com>