patch release schedule: add expected EOL for 1.18, 1.17
As a part of WG LTS's wind down, we've discussed that given expected
release cadences that 1.18 and 1.17 are going to get a year of patch
support.  It's been noted though that the schedule remained ambiguous
for these two.  This commit adds an EOL target date in the month
of their release one year anniversary.

Signed-off-by: Tim Pepper <tpepper@vmware.com>