Update CI Signal Issue format to be more conducive to triage
This updates the documentation on openining issues for the CI Signal
team based on feedback from SIGs. The primary changes include
centralizing all information for opening issues to a single section, as
well as providing a decision tree that informs a team member whether
they should open one or multiple issues, and how each issue should be
named.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>