Update roadmap
We now split-up the deliverables to reflect the state better (WIP vs
TODO). This also means that we outline if no enhancement is available
for a certain item.

We also move "Distribute the load of Kubernetes artifacts between
vendors (Consumable)" to done since this is mostly in the scope of K8S
infra now and we supported the initial phase.

We also move "Simplify CVE process for release management (Secure)" to
done because there is just a single PR in the scope of Security Response
Committee which has to be merged:
https://github.com/kubernetes/committee-security-response/issues/63

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>